refactor(bdd): split oversized xbrlkit-bdd-steps module into submodules
Split the monolithic 1625-line lib.rs into logical submodules:
- types.rs: Core types (Step, World, context structs) and core functions
- given.rs: Given step handlers (493 lines)
- when.rs: When step handlers (388 lines)
- then.rs: Then step handlers and assertions (520 lines)
- utils.rs: Helper functions (76 lines)
- lib.rs: Minimal module re-exports (18 lines)

This refactoring:
- Improves maintainability by separating concerns
- Makes the codebase more navigable
- Allows independent evolution of step handler categories
- Reduces cognitive load when working on specific step types

Closes #155
