+++ crates/xbrlkit-bdd-steps/src/lib.rs
@@ -1,1625 +1,18 @@
 //! Minimal step execution for the active BDD slices.
-
-use anyhow::Context;
-use dimensional_rules::validate_context_dimensions;
-use scenario_contract::{BundleManifest, FeatureGrid, ScenarioRecord};
-use scenario_runner::{
-    ScenarioExecution, assert_scenario_outcome, ensure_ixds_member_count,
-    ensure_report_concept_set, ensure_report_contains_rule, ensure_report_does_not_contain_rule,
-    ensure_report_fact_count, ensure_report_has_no_error_findings,
-    ensure_taxonomy_resolution_resolves_at_least, ensure_taxonomy_resolution_succeeds,
-    execute_scenario, write_execution_receipts,
-};
-use std::path::PathBuf;
-use taxonomy_dimensions::{Dimension, DimensionTaxonomy, Domain, DomainMember};
-use xbrl_contexts::{DimensionMember, DimensionalContainer, EntityIdentifier, Period};
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Step {
-    pub text: String,
-    pub table: Vec<Vec<String>>,
-}
-
-#[derive(Debug, Clone)]
-pub struct World {
-    pub repo_root: PathBuf,
-    pub grid: FeatureGrid,
-    pub profile_id: Option<String>,
-    pub fixture_dirs: Vec<PathBuf>,
-    pub execution: Option<ScenarioExecution>,
-    pub dimension_context: DimensionContext,
-    pub context_completeness_context: ContextCompletenessContext,
-    pub streaming_context: StreamingContext,
-    pub taxonomy_loader_context: TaxonomyLoaderContext,
-    pub bundle_manifest: Option<BundleManifest>,
-    pub validation_receipt: Option<receipt_types::Receipt>,
-    pub sensor_report: Option<serde_json::Value>,
-    pub filing_manifest: Option<edgar_attachments::FilingManifest>,
-    pub filing_receipt: Option<receipt_types::Receipt>,
-    pub compiled_grid: Option<FeatureGrid>,
-    pub cli_output: Option<String>,
-    pub cli_json_output: Option<serde_json::Value>,
-    pub cli_exit_code: Option<i32>,
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct DimensionContext {
-    pub dimension: Option<String>,
-    pub member: Option<String>,
-    pub concept: Option<String>,
-    pub required_dimension: Option<String>,
-    pub validation_findings: Vec<String>,
-    pub typed_value_type: Option<String>, // Added for typed value validation
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct ContextCompletenessContext {
-    pub contexts: Vec<xbrl_contexts::Context>,
-    pub facts: Vec<xbrl_report_types::Fact>,
-    pub findings: Vec<xbrl_report_types::ValidationFinding>,
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct StreamingContext {
-    pub file_size_mb: Option<f64>,
-    pub fact_count: Option<usize>,
-    pub memory_peak_mb: Option<f64>,
-    pub facts_processed: Vec<xbrl_stream::StreamingFact>,
-    pub contexts_collected: Vec<xbrl_stream::StreamingContext>,
-    pub units_collected: Vec<xbrl_stream::StreamingUnit>,
-    pub use_streaming: bool,
-    pub missing_context_refs: Vec<String>,
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct TaxonomyLoaderContext {
-    pub loader: Option<taxonomy_loader::TaxonomyLoader>,
-    pub taxonomy: Option<DimensionTaxonomy>,
-    pub cache_dir: Option<PathBuf>,
-    pub schema_path: Option<String>,
-    pub loaded: bool,
-}
-
-impl World {
-    #[must_use]
-    pub fn new(repo_root: PathBuf, grid: FeatureGrid) -> Self {
-        Self {
-            repo_root,
-            grid,
-            profile_id: None,
-            fixture_dirs: Vec::new(),
-            execution: None,
-            dimension_context: DimensionContext::default(),
-            context_completeness_context: ContextCompletenessContext::default(),
-            streaming_context: StreamingContext::default(),
-            taxonomy_loader_context: TaxonomyLoaderContext::default(),
-            bundle_manifest: None,
-            validation_receipt: None,
-            sensor_report: None,
-            filing_manifest: None,
-            filing_receipt: None,
-            compiled_grid: None,
-            cli_output: None,
-            cli_json_output: None,
-            cli_exit_code: None,
-        }
-    }
-}
-
-pub fn run_scenario(
-    world: &mut World,
-    scenario: &ScenarioRecord,
-    steps: &[Step],
-) -> anyhow::Result<()> {
-    if !world
-        .grid
-        .scenarios
-        .iter()
-        .any(|candidate| candidate.scenario_id == scenario.scenario_id)
-    {
-        anyhow::bail!(
-            "scenario {} is not present in the feature grid",
-            scenario.scenario_id
-        );
-    }
-
-    for step in steps {
-        run_step(world, scenario, step)?;
-    }
-
-    // Some scenarios (like dimension validation) validate via step assertions
-    // and don't set execution. Skip scenario-level outcome check in that case.
-    if let Some(execution) = world.execution.as_ref() {
-        assert_scenario_outcome(scenario, execution)?;
-    }
-
-    Ok(())
-}
-
-fn run_step(world: &mut World, scenario: &ScenarioRecord, step: &Step) -> anyhow::Result<()> {
-    if handle_given(world, scenario, step)? {
-        return Ok(());
-    }
-    if handle_when(world, scenario, step)? {
-        return Ok(());
-    }
-    handle_then(world, step)
-}
-
-fn execution(world: &World) -> anyhow::Result<&ScenarioExecution> {
-    world
-        .execution
-        .as_ref()
-        .context("scenario step requires a prior execution")
-}
-
-fn assert_declared_inputs_match(world: &World, scenario: &ScenarioRecord) -> anyhow::Result<()> {
-    if let Some(profile_id) = &world.profile_id {
-        if scenario.profile_pack.as_deref() != Some(profile_id.as_str()) {
-            anyhow::bail!("declared profile pack does not match scenario metadata");
-        }
-    }
-
-    if !world.fixture_dirs.is_empty() {
-        let declared = world
-            .fixture_dirs
-            .iter()
-            .map(|path| {
-                path.strip_prefix(world.repo_root.join("fixtures"))
-                    .expect("fixture path under repo root")
-                    .to_string_lossy()
-                    .replace('\\', "/")
-            })
-            .collect::<Vec<_>>();
-        if declared != scenario.fixtures {
-            anyhow::bail!("declared fixture directories do not match scenario metadata");
-        }
-    }
-
-    Ok(())
-}
-
-#[allow(clippy::too_many_lines)]
-fn handle_given(world: &mut World, scenario: &ScenarioRecord, step: &Step) -> anyhow::Result<bool> {
-    if let Some(profile_id) = step.text.strip_prefix("the profile pack \"") {
-        let profile_id = profile_id.trim_end_matches('"').to_string();
-        if scenario.profile_pack.as_deref() != Some(profile_id.as_str()) {
-            anyhow::bail!(
-                "feature file profile pack {profile_id} does not match scenario metadata"
-            );
-        }
-        world.profile_id = Some(profile_id);
-        return Ok(true);
-    }
-
-    if let Some(fixture) = step
-        .text
-        .strip_prefix("the fixture directory \"")
-        .or_else(|| step.text.strip_prefix("the fixture \""))
-    {
-        let fixture = fixture.trim_end_matches('"');
-        if !scenario
-            .fixtures
-            .iter()
-            .any(|candidate| candidate == fixture)
-        {
-            anyhow::bail!("feature file fixture {fixture} does not match scenario metadata");
-        }
-        world
-            .fixture_dirs
-            .push(world.repo_root.join("fixtures").join(fixture));
-        return Ok(true);
-    }
-
-    // Dimension-related Given steps
-    if step.text == "the taxonomy has dimension definitions" {
-        // Stub - implement actual taxonomy loading when needed
-        return Ok(true);
-    }
-
-    if step.text == "the taxonomy has domain hierarchies" {
-        // Stub - implement actual taxonomy loading when needed
-        return Ok(true);
-    }
-
-    if step.text == "the taxonomy has hypercube definitions" {
-        // Stub - implement actual taxonomy loading when needed
-        return Ok(true);
-    }
-
-    if let Some(dimension) = step.text.strip_prefix("a context with dimension \"") {
-        world.dimension_context.dimension = Some(dimension.trim_end_matches('"').to_string());
-        return Ok(true);
-    }
-
-    if let Some(dimension) = step
-        .text
-        .strip_prefix("a context with unknown dimension \"")
-    {
-        world.dimension_context.dimension = Some(dimension.trim_end_matches('"').to_string());
-        return Ok(true);
-    }
-
-    if let Some(member) = step.text.strip_prefix("the member \"") {
-        world.dimension_context.member = Some(member.trim_end_matches('"').to_string());
-        return Ok(true);
-    }
-
-    if let Some(member) = step.text.strip_prefix("an invalid member \"") {
-        world.dimension_context.member = Some(member.trim_end_matches('"').to_string());
-        return Ok(true);
-    }
-
-    if let Some(concept) = step.text.strip_prefix("a fact for concept \"") {
-        world.dimension_context.concept = Some(concept.trim_end_matches('"').to_string());
-        return Ok(true);
-    }
-
-    if let Some(dimension) = step.text.strip_prefix("the concept requires dimension \"") {
-        world.dimension_context.required_dimension =
-            Some(dimension.trim_end_matches('"').to_string());
-        return Ok(true);
-    }
-
-    if step.text == "a context without that dimension" {
-        // Ensure dimension is not set (or clear it)
-        world.dimension_context.dimension = None;
-        return Ok(true);
-    }
-
-    // Typed dimension Given steps
-    if let Some(dimension) = step.text.strip_prefix("a context with typed dimension \"") {
-        let rest = dimension.trim_end_matches('"');
-        // Handle "dim:Axis" of type "xs:type" format
-        if let Some((dim, type_part)) = rest.split_once("\" of type \"") {
-            world.dimension_context.dimension = Some(dim.to_string());
-            world.dimension_context.typed_value_type = Some(type_part.to_string());
-        } else {
-            world.dimension_context.dimension = Some(rest.to_string());
-        }
-        return Ok(true);
-    }
-
-    if let Some(value) = step.text.strip_prefix("the typed member value \"") {
-        world.dimension_context.member = Some(value.trim_end_matches('"').to_string());
-        return Ok(true);
-    }
-
-    // Bundle-related Given steps
-    if step.text == "the feature grid is compiled" {
-        // Grid is already loaded in World, just verify it's not empty
-        if world.grid.scenarios.is_empty() {
-            anyhow::bail!("feature grid is empty");
-        }
-        return Ok(true);
-    }
-
-    // Feature grid Given steps
-    if step.text == "the repo has feature sidecars" {
-        // Check that at least one .meta.yaml file exists in specs/features
-        let features_root = world.repo_root.join("specs/features");
-        let has_sidecars = walkdir::WalkDir::new(&features_root)
-            .into_iter()
-            .filter_map(Result::ok)
-            .any(|entry: walkdir::DirEntry| {
-                let path = entry.path();
-                path.extension().is_some_and(|ext| ext == "yaml")
-                    && path
-                        .file_name()
-                        .and_then(|n: &std::ffi::OsStr| n.to_str())
-                        .is_some_and(|n: &str| n.ends_with(".meta.yaml"))
-            });
-        if !has_sidecars {
-            anyhow::bail!("no feature sidecars found in {}", features_root.display());
-        }
-        return Ok(true);
-    }
-
-    // Cockpit pack Given steps
-    if step.text == "a validation report receipt" {
-        world.validation_receipt = Some(receipt_types::Receipt::new(
-            "validation.report",
-            "synthetic-subject",
-            receipt_types::RunResult::Success,
-        ));
-        return Ok(true);
-    }
-
-    // CLI Given steps
-    if step.text == "a SEC profile is configured" {
-        world.profile_id = Some("sec/efm-77/opco".to_string());
-        return Ok(true);
-    }
-
-    // Alpha check Given steps
-    if step.text == "the active alpha scenarios are implemented" {
-        // Parse feature files to verify scenarios with @alpha-active tag exist
-        let features_root = world.repo_root.join("specs/features");
-        let mut has_alpha_scenarios = false;
-
-        for entry in walkdir::WalkDir::new(&features_root)
-            .into_iter()
-            .filter_map(Result::ok)
-        {
-            if !entry.file_type().is_file() {
-                continue;
-            }
-            let path = entry.path();
-            if path.extension().is_none_or(|ext| ext != "feature") {
-                continue;
-            }
-
-            if let Ok(content) = std::fs::read_to_string(path) {
-                // Check for @alpha-active tag in the file
-                if content.contains("@alpha-active") {
-                    has_alpha_scenarios = true;
-                    break;
-                }
-            }
-        }
-
-        if !has_alpha_scenarios {
-            anyhow::bail!("no scenarios with @alpha-active tag found in feature files");
-        }
-        return Ok(true);
-    }
-
-    // Context completeness Given steps
-    if step.text.starts_with("an XBRL report with context ") {
-        // Parse contexts from the step text
-        // Format: "an XBRL report with context \"ctx-1\"" or "an XBRL report with contexts \"ctx-1\" and \"ctx-2\""
-        let text = &step.text;
-        let contexts: Vec<String> = text
-            .split('"')
-            .enumerate()
-            .filter(|(i, _)| i % 2 == 1)
-            .map(|(_, s)| s.to_string())
-            .collect();
-
-        for ctx_id in contexts {
-            let context = xbrl_contexts::Context {
-                id: xbrl_contexts::normalize_context_id(&ctx_id),
-                entity: EntityIdentifier {
-                    scheme: "http://www.sec.gov/CIK".to_string(),
-                    value: "0000320193".to_string(),
-                },
-                entity_segment: None,
-                period: Period::Instant("2024-12-31".to_string()),
-                scenario: None,
-            };
-            world.context_completeness_context.contexts.push(context);
-        }
-        return Ok(true);
-    }
-
-    if step.text.starts_with("a fact referencing concept ") {
-        // Parse: "a fact referencing concept \"us-gaap:Revenue\" with context \"ctx-1\""
-        let text = &step.text;
-        if let Some(concept_start) = text.find('\"') {
-            let concept_end = text[concept_start + 1..]
-                .find('\"')
-                .map(|i| concept_start + 1 + i);
-            if let Some(concept_end) = concept_end {
-                let concept = &text[concept_start + 1..concept_end];
-                if let Some(ctx_start) = text[concept_end + 1..].find('\"') {
-                    let ctx_start = concept_end + 1 + ctx_start;
-                    let ctx_end = text[ctx_start + 1..].find('\"').map(|i| ctx_start + 1 + i);
-                    if let Some(ctx_end) = ctx_end {
-                        let context_ref = &text[ctx_start + 1..ctx_end];
-                        let fact = xbrl_report_types::Fact {
-                            concept: concept.to_string(),
-                            context_ref: context_ref.to_string(),
-                            unit_ref: None,
-                            decimals: None,
-                            value: "1000".to_string(),
-                            member: String::new(),
-                        };
-                        world.context_completeness_context.facts.push(fact);
-                        return Ok(true);
-                    }
-                }
-            }
-        }
-        anyhow::bail!("invalid fact specification: {}", step.text);
-    }
-
-    if step.text.starts_with("facts referencing concepts ") {
-        // Parse: "facts referencing concepts \"us-gaap:Revenue\" and \"us-gaap:Assets\" with contexts \"ctx-1\" and \"ctx-2\""
-        // For simplicity, we'll create facts for each concept-context pair
-        // Parse all quoted strings
-        let quoted: Vec<String> = step
-            .text
-            .split('\"')
-            .enumerate()
-            .filter(|(i, _)| i % 2 == 1)
-            .map(|(_, s)| s.to_string())
-            .collect();
-
-        if quoted.len() >= 2 {
-            // First half are concepts, second half are contexts
-            let mid = quoted.len() / 2;
-            let concepts = &quoted[..mid];
-            let contexts = &quoted[mid..];
-
-            for (i, concept) in concepts.iter().enumerate() {
-                let context_ref = contexts
-                    .get(i)
-                    .or_else(|| contexts.first())
-                    .map_or("ctx-1", String::as_str);
-                let fact = xbrl_report_types::Fact {
-                    concept: concept.clone(),
-                    context_ref: context_ref.to_string(),
-                    unit_ref: None,
-                    decimals: None,
-                    value: "1000".to_string(),
-                    member: String::new(),
-                };
-                world.context_completeness_context.facts.push(fact);
-            }
-        }
-        return Ok(true);
-    }
-
-    // Decimal precision Given steps
-    if step.text.starts_with("a numeric fact with value ") {
-        // Clear previous state for standalone decimal precision scenarios
-        world.context_completeness_context.facts.clear();
-        world.context_completeness_context.contexts.clear();
-        world.context_completeness_context.findings.clear();
-
-        // Parse: "a numeric fact with value "1234.56" and decimals "INF""
-        let quoted: Vec<String> = step
-            .text
-            .split('"')
-            .enumerate()
-            .filter(|(i, _)| i % 2 == 1)
-            .map(|(_, s)| s.to_string())
-            .collect();
-
-        if quoted.len() >= 2 {
-            let value = &quoted[0];
-            let decimals = &quoted[1];
-            let fact = xbrl_report_types::Fact {
-                concept: "us-gaap:TestConcept".to_string(),
-                context_ref: "ctx-1".to_string(),
-                unit_ref: Some("usd".to_string()),
-                decimals: Some(decimals.clone()),
-                value: value.clone(),
-                member: String::new(),
-            };
-            world.context_completeness_context.facts.push(fact);
-            return Ok(true);
-        }
-        anyhow::bail!("invalid numeric fact specification: {}", step.text);
-    }
-
-    // Streaming parser Given steps
-    if step.text == "the xbrl-stream crate is available" {
-        // Verify the crate exists and can be used
-        world.streaming_context.use_streaming = true;
-        return Ok(true);
-    }
-
-    if step.text.starts_with("an XBRL filing larger than ") {
-        // Parse: "an XBRL filing larger than 100MB"
-        let mb_str = step
-            .text
-            .strip_prefix("an XBRL filing larger than ")
-            .and_then(|s| s.strip_suffix("MB"))
-            .or_else(|| {
-                step.text
-                    .strip_prefix("an XBRL filing larger than ")
-                    .and_then(|s| s.strip_suffix("mb"))
-            });
-        if let Some(mb) = mb_str {
-            world.streaming_context.file_size_mb = Some(mb.parse().unwrap_or(100.0));
-            return Ok(true);
-        }
-        anyhow::bail!("invalid file size specification: {}", step.text);
-    }
-
-    if step.text.starts_with("an XBRL filing smaller than ") {
-        // Parse: "an XBRL filing smaller than 10MB"
-        let mb_str = step
-            .text
-            .strip_prefix("an XBRL filing smaller than ")
-            .and_then(|s| s.strip_suffix("MB"))
-            .or_else(|| {
-                step.text
-                    .strip_prefix("an XBRL filing smaller than ")
-                    .and_then(|s| s.strip_suffix("mb"))
-            });
-        if let Some(mb) = mb_str {
-            world.streaming_context.file_size_mb = Some(mb.parse().unwrap_or(10.0));
-            world.streaming_context.use_streaming = true; // Always available as option
-            return Ok(true);
-        }
-        anyhow::bail!("invalid file size specification: {}", step.text);
-    }
-
-    if step.text.starts_with("a large XBRL filing with ") {
-        // Parse: "a large XBRL filing with 1000+ facts"
-        if let Some(facts_str) = step.text.strip_prefix("a large XBRL filing with ") {
-            let facts = facts_str
-                .split('+')
-                .next()
-                .and_then(|s| s.trim().parse().ok())
-                .unwrap_or(1000);
-            world.streaming_context.fact_count = Some(facts);
-            world.streaming_context.file_size_mb = Some(50.0); // Assume large
-            return Ok(true);
-        }
-        anyhow::bail!("invalid fact count specification: {}", step.text);
-    }
-
-    if step.text == "some facts reference non-existent contexts" {
-        // Mark that we'll simulate missing context refs
-        world.streaming_context.missing_context_refs = vec!["missing-ctx-1".to_string()];
-        return Ok(true);
-    }
-
-    if step.text == "a streaming parser with a custom handler" {
-        world.streaming_context.use_streaming = true;
-        world.streaming_context.facts_processed.clear();
-        world.streaming_context.contexts_collected.clear();
-        world.streaming_context.units_collected.clear();
-        return Ok(true);
-    }
-
-    // Taxonomy loader Given steps
-    if step.text == "the taxonomy loader is available" {
-        world.taxonomy_loader_context.loader = Some(taxonomy_loader::TaxonomyLoader::new());
-        return Ok(true);
-    }
-
-    if step.text == "a taxonomy schema with dimension elements" {
-        // Use a fixture path or create synthetic schema
-        world.taxonomy_loader_context.schema_path =
-            Some("fixtures/synthetic/taxonomy/standard-location-01/schema.xsd".to_string());
-        world.taxonomy_loader_context.loader = Some(taxonomy_loader::TaxonomyLoader::new());
-        return Ok(true);
-    }
-
-    if step.text == "a taxonomy definition linkbase with domain members" {
-        world.taxonomy_loader_context.schema_path =
-            Some("fixtures/synthetic/taxonomy/standard-location-01/schema.xsd".to_string());
-        world.taxonomy_loader_context.loader = Some(taxonomy_loader::TaxonomyLoader::new());
-        return Ok(true);
-    }
-
-    if step.text == "a taxonomy with typed dimensions" {
-        world.taxonomy_loader_context.schema_path =
-            Some("fixtures/synthetic/taxonomy/standard-location-01/schema.xsd".to_string());
-        world.taxonomy_loader_context.loader = Some(taxonomy_loader::TaxonomyLoader::new());
-        return Ok(true);
-    }
-
-    if step.text == "a taxonomy with hypercube elements" {
-        world.taxonomy_loader_context.schema_path =
-            Some("fixtures/synthetic/taxonomy/standard-location-01/schema.xsd".to_string());
-        world.taxonomy_loader_context.loader = Some(taxonomy_loader::TaxonomyLoader::new());
-        return Ok(true);
-    }
-
-    if step.text == "a taxonomy URL to load" {
-        // Use a synthetic path that doesn't exist - triggers synthetic taxonomy creation
-        world.taxonomy_loader_context.schema_path =
-            Some("fixtures/synthetic/taxonomy/url-test/schema.xsd".to_string());
-        return Ok(true);
-    }
-
-    if step.text == "a cache directory is configured" {
-        let cache_dir = std::env::temp_dir().join("xbrlkit_taxonomy_cache");
-        // Create the cache directory so it exists for the Then step
-        let _ = std::fs::create_dir_all(&cache_dir);
-        world.taxonomy_loader_context.cache_dir = Some(cache_dir.clone());
-        world.taxonomy_loader_context.loader =
-            Some(taxonomy_loader::TaxonomyLoader::with_cache_dir(&cache_dir));
-        return Ok(true);
-    }
-
-    if step.text == "a taxonomy schema that imports another schema" {
-        world.taxonomy_loader_context.schema_path =
-            Some("fixtures/synthetic/taxonomy/standard-location-01/schema.xsd".to_string());
-        world.taxonomy_loader_context.loader = Some(taxonomy_loader::TaxonomyLoader::new());
-        return Ok(true);
-    }
-
-    if step.text == "a loaded taxonomy with dimension definitions" {
-        // Simulate loading a taxonomy with dimensions
-        let mut taxonomy = DimensionTaxonomy::new();
-
-        // Add a domain
-        let mut domain = Domain::new("us-gaap:ScenarioDomain");
-        domain.add_member(DomainMember {
-            qname: "us-gaap:ScenarioActualMember".to_string(),
-            parent: None,
-            order: 1,
-            label: None,
-        });
-        domain.add_member(DomainMember {
-            qname: "us-gaap:ScenarioForecastMember".to_string(),
-            parent: None,
-            order: 2,
-            label: None,
-        });
-        taxonomy.add_domain(domain);
-
-        // Add an explicit dimension
-        taxonomy.add_dimension(Dimension::Explicit {
-            qname: "us-gaap:StatementScenarioAxis".to_string(),
-            default_domain: Some("us-gaap:ScenarioDomain".to_string()),
-            required: false,
-        });
-        taxonomy.dimension_domains.insert(
-            "us-gaap:StatementScenarioAxis".to_string(),
-            "us-gaap:ScenarioDomain".to_string(),
-        );
-
-        world.taxonomy_loader_context.taxonomy = Some(taxonomy);
-        world.taxonomy_loader_context.loaded = true;
-        return Ok(true);
-    }
-
-    Ok(false)
-}
-
-#[allow(clippy::too_many_lines)]
-fn handle_when(world: &mut World, scenario: &ScenarioRecord, step: &Step) -> anyhow::Result<bool> {
-    if matches!(
-        step.text.as_str(),
-        "I validate the filing" | "I validate duplicate facts" | "I resolve the DTS"
-    ) {
-        assert_declared_inputs_match(world, scenario)?;
-        let execution = execute_scenario(&world.repo_root, scenario)?;
-        write_execution_receipts(&world.repo_root, &execution)?;
-        world.execution = Some(execution);
-        return Ok(true);
-    }
-
-    if step.text == "I export the canonical report to JSON" {
-        assert_declared_inputs_match(world, scenario)?;
-        let execution = execute_scenario(&world.repo_root, scenario)?;
-        write_execution_receipts(&world.repo_root, &execution)?;
-        world.execution = Some(execution);
-        return Ok(true);
-    }
-
-    // Feature grid When steps
-    if step.text == "I compile the feature grid" {
-        world.compiled_grid = Some(xbrlkit_feature_grid::compile(&world.repo_root)?);
-        return Ok(true);
-    }
-
-    // Dimension-related When steps
-    if step.text == "I validate the dimension-member pair" {
-        world.dimension_context.validation_findings.clear();
-
-        let dimension = world.dimension_context.dimension.as_deref().unwrap_or("");
-        let member = world.dimension_context.member.as_deref().unwrap_or("");
-
-        // Build minimal taxonomy with StatementScenarioAxis
-        let mut taxonomy = DimensionTaxonomy::new();
-        let mut scenario_domain = Domain::new("us-gaap:ScenarioDomain");
-        scenario_domain.add_member(DomainMember {
-            qname: "us-gaap:ScenarioActualMember".to_string(),
-            parent: None,
-            order: 1,
-            label: None,
-        });
-        scenario_domain.add_member(DomainMember {
-            qname: "us-gaap:ScenarioForecastMember".to_string(),
-            parent: None,
-            order: 2,
-            label: None,
-        });
-        taxonomy.add_domain(scenario_domain);
-
-        taxonomy.add_dimension(Dimension::Explicit {
-            qname: "us-gaap:StatementScenarioAxis".to_string(),
-            default_domain: Some("us-gaap:ScenarioDomain".to_string()),
-            required: false,
-        });
-        taxonomy.dimension_domains.insert(
-            "us-gaap:StatementScenarioAxis".to_string(),
-            "us-gaap:ScenarioDomain".to_string(),
-        );
-
-        // Build context with dimensional information in scenario
-        let mut context = xbrl_contexts::Context {
-            id: "test-context".to_string(),
-            entity: EntityIdentifier {
-                scheme: "http://www.sec.gov/CIK".to_string(),
-                value: "0001234567".to_string(),
-            },
-            period: Period::Duration {
-                start: "2024-01-01".to_string(),
-                end: "2024-12-31".to_string(),
-            },
-            entity_segment: None,
-            scenario: None,
-        };
-
-        // Add dimensional member if specified
-        if !dimension.is_empty() && !member.is_empty() {
-            context.scenario = Some(DimensionalContainer {
-                dimensions: vec![DimensionMember {
-                    dimension: dimension.to_string(),
-                    member: member.to_string(),
-                    is_typed: false,
-                    typed_value: None,
-                }],
-                raw_xml: None,
-            });
-        }
-
-        // Validate
-        // Note: concept is empty here because this step validates dimension-member
-        // pairs independently of any concept. Required dimension checking happens
-        // in "I validate the fact dimensions" which provides a concept.
-        let result = validate_context_dimensions(&context, "", &taxonomy);
-        for finding in result.findings {
-            world
-                .dimension_context
-                .validation_findings
-                .push(finding.rule_id.clone());
-        }
-
-        return Ok(true);
-    }
-
-    if step.text == "I validate the fact dimensions" {
-        world.dimension_context.validation_findings.clear();
-
-        let _concept = world.dimension_context.concept.as_deref().unwrap_or("");
-        let has_dimension = world.dimension_context.dimension.is_some();
-        let required_dim = world.dimension_context.required_dimension.clone();
-
-        // Check for missing required dimension
-        if let Some(_req_dim) = required_dim {
-            if !has_dimension {
-                world
-                    .dimension_context
-                    .validation_findings
-                    .push("XBRL.DIMENSION.MISSING_REQUIRED".to_string());
-            }
-        }
-
-        return Ok(true);
-    }
-
-    if step.text == "I validate the typed dimension value" {
-        world.dimension_context.validation_findings.clear();
-
-        let dimension = world.dimension_context.dimension.as_deref().unwrap_or("");
-        let value = world.dimension_context.member.as_deref().unwrap_or("");
-        let value_type = world
-            .dimension_context
-            .typed_value_type
-            .as_deref()
-            .unwrap_or("xs:string");
-
-        // Build taxonomy with typed dimension
-        let mut taxonomy = DimensionTaxonomy::new();
-        taxonomy.add_dimension(Dimension::Typed {
-            qname: dimension.to_string(),
-            value_type: value_type.to_string(),
-            required: false,
-        });
-
-        // Build context with typed dimension
-        let context = xbrl_contexts::Context {
-            id: "test-context".to_string(),
-            entity: EntityIdentifier {
-                scheme: "http://www.sec.gov/CIK".to_string(),
-                value: "0001234567".to_string(),
-            },
-            period: Period::Duration {
-                start: "2024-01-01".to_string(),
-                end: "2024-12-31".to_string(),
-            },
-            entity_segment: None,
-            scenario: Some(DimensionalContainer {
-                dimensions: vec![DimensionMember {
-                    dimension: dimension.to_string(),
-                    member: value.to_string(),
-                    is_typed: true,
-                    typed_value: Some(value.to_string()),
-                }],
-                raw_xml: None,
-            }),
-        };
-
-        // Validate
-        let result = validate_context_dimensions(&context, "", &taxonomy);
-        for finding in result.findings {
-            world
-                .dimension_context
-                .validation_findings
-                .push(finding.rule_id.clone());
-        }
-
-        return Ok(true);
-    }
-
-    // Bundle-related When steps
-    if let Some(selector) = step.text.strip_prefix("I bundle the selector \"") {
-        let selector = selector.trim_end_matches('"').to_string();
-        let scenarios = select_matching_scenarios(&world.grid, &selector);
-        world.bundle_manifest = Some(BundleManifest {
-            selector,
-            scenarios,
-        });
-        return Ok(true);
-    }
-
-    if step.text == "I build the filing manifest" {
-        let fixture_dir = world.fixture_dirs.first().context("no fixture loaded")?;
-        let submission_path = fixture_dir.join("submission.txt");
-        let submission = std::fs::read_to_string(&submission_path).with_context(|| {
-            format!(
-                "failed to read submission.txt from {}",
-                fixture_dir.display()
-            )
-        })?;
-        let (manifest, receipt) = filing_load::load_from_submission(&submission);
-        world.filing_manifest = Some(manifest);
-        world.filing_receipt = Some(receipt);
-        return Ok(true);
-    }
-
-    // Cockpit pack When steps
-    if step.text == "I package the receipt for cockpit" {
-        let receipt = world
-            .validation_receipt
-            .as_ref()
-            .context("packaging requires a validation receipt")?;
-        world.sensor_report = Some(cockpit_export::to_sensor_report("xbrlkit", receipt));
-        return Ok(true);
-    }
-
-    // CLI When steps
-    if step.text == "I run describe-profile --json" {
-        let profile_id = world
-            .profile_id
-            .as_ref()
-            .context("profile must be configured")?;
-        let profile = sec_profile_types::load_profile_from_workspace(&world.repo_root, profile_id)?;
-        let json_output =
-            serde_json::to_string_pretty(&profile).context("serializing profile to JSON")?;
-        world.cli_output = Some(json_output);
-        return Ok(true);
-    }
-
-    // Alpha check When steps
-    if step.text == "I run the alpha readiness gate" {
-        // Instead of running bdd (which causes recursion), just verify the grid can be loaded
-        // The Given step already verified @alpha-active scenarios exist
-        // This step just confirms the test infrastructure is ready
-        world.cli_output = Some("alpha scenarios verified".to_string());
-        world.cli_exit_code = Some(0);
-        return Ok(true);
-    }
-
-    // Context completeness When steps
-    if step.text == "context completeness validation runs" {
-        // Build ContextSet from contexts
-        let mut context_set = xbrl_contexts::ContextSet::new();
-        for ctx in &world.context_completeness_context.contexts {
-            context_set.insert(ctx.clone());
-        }
-        // Run validation
-        let findings = context_completeness::validate_context_completeness(
-            &world.context_completeness_context.facts,
-            &context_set,
-        );
-        world.context_completeness_context.findings = findings;
-        return Ok(true);
-    }
-
-    // Decimal precision When steps
-    if step.text == "decimal precision validation is performed" {
-        let findings =
-            numeric_rules::validate_decimal_precision(&world.context_completeness_context.facts);
-        world.context_completeness_context.findings = findings;
-        return Ok(true);
-    }
-
-    // Streaming parser When steps
-    if step.text == "I validate it using the streaming parser" {
-        // Simulate streaming validation - in real implementation this would
-        // use xbrl_stream to parse a large file
-        world.streaming_context.memory_peak_mb = Some(45.0); // Simulated under 50MB
-        world.streaming_context.facts_processed = vec![xbrl_stream::StreamingFact {
-            concept: "us-gaap:Revenue".to_string(),
-            context_ref: "ctx-1".to_string(),
-            unit_ref: Some("usd".to_string()),
-            decimals: Some("-3".to_string()),
-            value: "12345000".to_string(),
-        }];
-        return Ok(true);
-    }
-
-    if step.text == "I check if streaming is needed" {
-        // Determine if streaming should be recommended based on file size
-        // Streaming is always available as an option, but recommended for large files
-        let _size = world.streaming_context.file_size_mb.unwrap_or(0.0);
-        // use_streaming remains true (set in Given step) to indicate availability
-        return Ok(true);
-    }
-
-    if step.text == "I run streaming context validation" {
-        // Simulate streaming context validation
-        world.streaming_context.facts_processed = vec![
-            xbrl_stream::StreamingFact {
-                concept: "us-gaap:Revenue".to_string(),
-                context_ref: "ctx-1".to_string(),
-                unit_ref: Some("usd".to_string()),
-                decimals: Some("-3".to_string()),
-                value: "1000".to_string(),
-            },
-            xbrl_stream::StreamingFact {
-                concept: "us-gaap:Assets".to_string(),
-                context_ref: "missing-ctx-1".to_string(),
-                unit_ref: Some("usd".to_string()),
-                decimals: Some("-3".to_string()),
-                value: "2000".to_string(),
-            },
-        ];
-        world.streaming_context.contexts_collected = vec![xbrl_stream::StreamingContext {
-            id: "ctx-1".to_string(),
-            entity_scheme: Some("http://www.sec.gov/CIK".to_string()),
-            entity_value: Some("0001234567".to_string()),
-            period: xbrl_stream::StreamingPeriod::Instant("2024-12-31".to_string()),
-        }];
-        // Detect missing context refs
-        let context_ids: std::collections::HashSet<_> = world
-            .streaming_context
-            .contexts_collected
-            .iter()
-            .map(|c| c.id.clone())
-            .collect();
-        world.streaming_context.missing_context_refs = world
-            .streaming_context
-            .facts_processed
-            .iter()
-            .filter(|f| !context_ids.contains(&f.context_ref))
-            .map(|f| f.context_ref.clone())
-            .collect();
-        return Ok(true);
-    }
-
-    if step.text == "facts are encountered during parsing" {
-        // Simulate fact parsing with custom handler
-        world.streaming_context.facts_processed = vec![xbrl_stream::StreamingFact {
-            concept: "us-gaap:Revenue".to_string(),
-            context_ref: "ctx-1".to_string(),
-            unit_ref: Some("usd".to_string()),
-            decimals: Some("-3".to_string()),
-            value: "12345000".to_string(),
-        }];
-        world.streaming_context.contexts_collected = vec![xbrl_stream::StreamingContext {
-            id: "ctx-1".to_string(),
-            entity_scheme: Some("http://www.sec.gov/CIK".to_string()),
-            entity_value: Some("0001234567".to_string()),
-            period: xbrl_stream::StreamingPeriod::Instant("2024-12-31".to_string()),
-        }];
-        world.streaming_context.units_collected = vec![xbrl_stream::StreamingUnit {
-            id: "usd".to_string(),
-            measure: Some("iso4217:USD".to_string()),
-        }];
-        return Ok(true);
-    }
-
-    // Taxonomy loader When steps
-    if step.text == "I load the taxonomy" {
-        let loader = world
-            .taxonomy_loader_context
-            .loader
-            .take()
-            .context("taxonomy loader not initialized")?;
-        let schema_path = world
-            .taxonomy_loader_context
-            .schema_path
-            .clone()
-            .context("schema path not set")?;
-
-        // For synthetic test schemas that may not exist, create a minimal taxonomy
-        let taxonomy =
-            if schema_path.contains("fixtures/") && !std::path::Path::new(&schema_path).exists() {
-                // Create synthetic taxonomy for testing
-                create_synthetic_taxonomy()
-            } else {
-                loader.load(&schema_path)?
-            };
-
-        world.taxonomy_loader_context.taxonomy = Some(taxonomy);
-        world.taxonomy_loader_context.loaded = true;
-        return Ok(true);
-    }
-
-    Ok(false)
-}
-
-#[allow(clippy::too_many_lines)]
-fn handle_then(world: &mut World, step: &Step) -> anyhow::Result<()> {
-    // Dimension-related Then steps
-    if step.text == "the validation should pass" {
-        if !world.dimension_context.validation_findings.is_empty() {
-            anyhow::bail!(
-                "expected validation to pass but got findings: {:?}",
-                world.dimension_context.validation_findings
-            );
-        }
-        return Ok(());
-    }
-
-    if step.text == "the validation should fail" {
-        if world.dimension_context.validation_findings.is_empty() {
-            anyhow::bail!("expected validation to fail but no findings were reported");
-        }
-        return Ok(());
-    }
-
-    if step.text == "no findings should be reported" {
-        if !world.dimension_context.validation_findings.is_empty() {
-            anyhow::bail!(
-                "expected no findings but got: {:?}",
-                world.dimension_context.validation_findings
-            );
-        }
-        return Ok(());
-    }
-
-    if let Some(finding) = step.text.strip_prefix("an \"") {
-        let expected_finding = finding.trim_end_matches("\" finding should be reported");
-        if !world
-            .dimension_context
-            .validation_findings
-            .iter()
-            .any(|f| f == expected_finding)
-        {
-            anyhow::bail!(
-                "expected finding {} but got {:?}",
-                expected_finding,
-                world.dimension_context.validation_findings
-            );
-        }
-        return Ok(());
-    }
-
-    // Decimal precision Then steps
-    if step.text == "no validation errors are reported" {
-        if !world.context_completeness_context.findings.is_empty() {
-            anyhow::bail!(
-                "expected no validation errors but got: {:?}",
-                world.context_completeness_context.findings
-            );
-        }
-        return Ok(());
-    }
-
-    if let Some(error_type) = step.text.strip_prefix("validation error \"") {
-        let expected_error = error_type.trim_end_matches("\" is reported");
-        let has_error = world
-            .context_completeness_context
-            .findings
-            .iter()
-            .any(|f| f.rule_id.contains(expected_error) || f.message.contains(expected_error));
-        if !has_error {
-            anyhow::bail!(
-                "expected validation error '{}' but got: {:?}",
-                expected_error,
-                world.context_completeness_context.findings
-            );
-        }
-        return Ok(());
-    }
-
-    match step.text.as_str() {
-        "the validation report has no error findings" => {
-            ensure_report_has_no_error_findings(execution(world)?)
-        }
-        "the taxonomy resolution succeeds" => {
-            ensure_taxonomy_resolution_succeeds(execution(world)?)
-        }
-        "the concept set is:" => {
-            let expected = step
-                .table
-                .iter()
-                .filter_map(|row| row.first())
-                .map(String::as_str)
-                .collect::<Vec<_>>();
-            ensure_report_concept_set(execution(world)?, &expected)
-        }
-        "the export report receipt is emitted" => {
-            let execution = execution(world)?;
-            if execution.export_receipt.is_none() {
-                anyhow::bail!("export report receipt was not emitted");
-            }
-            Ok(())
-        }
-        "bundling fails because no scenario matches" => {
-            let manifest = world
-                .bundle_manifest
-                .as_ref()
-                .context("bundle step requires a prior bundle operation")?;
-            if !manifest.scenarios.is_empty() {
-                anyhow::bail!(
-                    "expected bundling to fail but found {} matching scenario(s)",
-                    manifest.scenarios.len()
-                );
-            }
-            Ok(())
-        }
-        "the sensor report is emitted" => {
-            if world.sensor_report.is_none() {
-                anyhow::bail!("sensor report was not emitted");
-            }
-            Ok(())
-        }
-        "the filing manifest receipt is emitted" => {
-            let receipt = world
-                .filing_receipt
-                .as_ref()
-                .context("filing manifest receipt was not emitted")?;
-            if receipt.kind != "filing.manifest" {
-                anyhow::bail!(
-                    "expected receipt kind 'filing.manifest', got '{}'",
-                    receipt.kind
-                );
-            }
-            Ok(())
-        }
-        // CLI Then steps
-        "the output is valid JSON" => {
-            let output = world
-                .cli_output
-                .clone()
-                .context("CLI output not captured")?;
-            let json_value: serde_json::Value =
-                serde_json::from_str(&output).context("CLI output is not valid JSON")?;
-            world.cli_json_output = Some(json_value);
-            Ok(())
-        }
-        "the profile contains required fields" => {
-            let json_value = world
-                .cli_json_output
-                .as_ref()
-                .context("JSON output not parsed")?;
-            let required_fields = [
-                "id",
-                "label",
-                "forms",
-                "enabled_rule_families",
-                "standard_taxonomy_uris",
-                "required_facts",
-            ];
-            for field in &required_fields {
-                if json_value.get(field).is_none() {
-                    anyhow::bail!("required field '{field}' is missing from profile output");
-                }
-            }
-            Ok(())
-        }
-        "the alpha readiness checks pass" => {
-            let exit_code = world
-                .cli_exit_code
-                .context("alpha readiness gate was not executed")?;
-            if exit_code != 0 {
-                let output = world.cli_output.as_deref().unwrap_or("no output captured");
-                anyhow::bail!(
-                    "alpha readiness gate failed with exit code {exit_code}\noutput:\n{output}"
-                );
-            }
-            Ok(())
-        }
-        _ => handle_parameterized_assertion(world, step),
-    }
-}
-
-#[allow(clippy::too_many_lines)]
-fn handle_parameterized_assertion(world: &World, step: &Step) -> anyhow::Result<()> {
-    if let Some(rule_id) = step
-        .text
-        .strip_prefix("the validation report contains rule \"")
-    {
-        let validation_run = execution(world)?
-            .validation_run
-            .as_ref()
-            .context("missing validation run")?;
-        return ensure_report_contains_rule(validation_run, rule_id.trim_end_matches('"'));
-    }
-
-    if let Some(rule_id) = step
-        .text
-        .strip_prefix("the validation report does not contain rule \"")
-    {
-        let validation_run = execution(world)?
-            .validation_run
-            .as_ref()
-            .context("missing validation run")?;
-        return ensure_report_does_not_contain_rule(validation_run, rule_id.trim_end_matches('"'));
-    }
-
-    if let Some(member_count) =
-        parse_count_suffix(&step.text, "the IXDS assembly receipt contains ", "member")
-    {
-        return ensure_ixds_member_count(execution(world)?, member_count);
-    }
-
-    if let Some(namespace_count) = parse_count_suffix(
-        &step.text,
-        "the taxonomy resolution resolves at least ",
-        "namespace",
-    ) {
-        return ensure_taxonomy_resolution_resolves_at_least(execution(world)?, namespace_count);
-    }
-
-    if let Some(fact_count) = parse_count_suffix(&step.text, "the report contains ", "fact") {
-        return ensure_report_fact_count(execution(world)?, fact_count);
-    }
-
-    // Bundle-related assertions
-    if let Some(scenario_id) = step
-        .text
-        .strip_prefix("the bundle manifest lists scenario \"")
-    {
-        let scenario_id = scenario_id.trim_end_matches('"');
-        let manifest = world
-            .bundle_manifest
-            .as_ref()
-            .context("bundle assertion requires a prior bundle operation")?;
-        if !manifest
-            .scenarios
-            .iter()
-            .any(|s| s.scenario_id == scenario_id)
-        {
-            anyhow::bail!(
-                "scenario {} not found in bundle manifest (contains {} scenario(s))",
-                scenario_id,
-                manifest.scenarios.len()
-            );
-        }
-        return Ok(());
-    }
-
-    // Feature grid assertions
-    if let Some(scenario_id) = step
-        .text
-        .strip_prefix("the feature grid contains scenario \"")
-    {
-        let scenario_id = scenario_id.trim_end_matches('"');
-        let grid = world
-            .compiled_grid
-            .as_ref()
-            .context("feature grid assertion requires a prior compile operation")?;
-        if !grid.scenarios.iter().any(|s| s.scenario_id == scenario_id) {
-            anyhow::bail!(
-                "scenario {} not found in feature grid (contains {} scenario(s))",
-                scenario_id,
-                grid.scenarios.len()
-            );
-        }
-        return Ok(());
-    }
-
-    // Context completeness Then steps
-    if let Some(context_ref) = step
-        .text
-        .strip_prefix("a context-missing error is reported for context \"")
-    {
-        let context_ref = context_ref.trim_end_matches('"');
-        let found = world
-            .context_completeness_context
-            .findings
-            .iter()
-            .any(|f| f.rule_id == "SEC-CONTEXT-001" && f.message.contains(context_ref));
-        if !found {
-            anyhow::bail!(
-                "expected context-missing error for '{}' but got findings: {:?}",
-                context_ref,
-                world.context_completeness_context.findings
-            );
-        }
-        return Ok(());
-    }
-
-    if step.text == "no context completeness findings are reported" {
-        if !world.context_completeness_context.findings.is_empty() {
-            anyhow::bail!(
-                "expected no findings but got: {:?}",
-                world.context_completeness_context.findings
-            );
-        }
-        return Ok(());
-    }
-
-    if let Some(count_str) = step
-        .text
-        .strip_prefix("context-missing errors are reported")
-    {
-        let expected_count: usize = count_str
-            .split_whitespace()
-            .next()
-            .and_then(|s| s.parse().ok())
-            .unwrap_or(1);
-        let actual_count = world
-            .context_completeness_context
-            .findings
-            .iter()
-            .filter(|f| f.rule_id == "SEC-CONTEXT-001")
-            .count();
-        if actual_count != expected_count {
-            anyhow::bail!(
-                "expected {} context-missing errors but got {}: {:?}",
-                expected_count,
-                actual_count,
-                world.context_completeness_context.findings
-            );
-        }
-        return Ok(());
-    }
-
-    if let Some(rule_id) = step.text.strip_prefix("the finding rule ID is \"") {
-        let rule_id = rule_id.trim_end_matches('"');
-        let found = world
-            .context_completeness_context
-            .findings
-            .iter()
-            .any(|f| f.rule_id == rule_id);
-        if !found {
-            anyhow::bail!(
-                "expected finding with rule ID '{}' but got: {:?}",
-                rule_id,
-                world.context_completeness_context.findings
-            );
-        }
-        return Ok(());
-    }
-
-    // Streaming parser Then steps
-    if step.text == "memory usage should stay under 50MB peak" {
-        let peak = world.streaming_context.memory_peak_mb.unwrap_or(f64::MAX);
-        if peak > 50.0 {
-            anyhow::bail!("memory usage was {peak}MB, expected under 50MB");
-        }
-        return Ok(());
-    }
-
-    if step.text == "all facts should be processed" {
-        if world.streaming_context.facts_processed.is_empty() {
-            anyhow::bail!("no facts were processed");
-        }
-        return Ok(());
-    }
-
-    if step.text == "context references should be validated" {
-        // Context refs were validated during streaming parse
-        return Ok(());
-    }
-
-    if step.text == "the DOM parser should be recommended" {
-        let size = world.streaming_context.file_size_mb.unwrap_or(0.0);
-        if size > 10.0 {
-            anyhow::bail!(
-                "DOM parser should be recommended for files under 10MB, but file is {size}MB"
-            );
-        }
-        return Ok(());
-    }
-
-    if step.text == "the streaming parser should be available as option" {
-        if !world.streaming_context.use_streaming {
-            anyhow::bail!("streaming parser should be available as an option");
-        }
-        return Ok(());
-    }
-
-    if step.text == "missing context references should be reported" {
-        if world.streaming_context.missing_context_refs.is_empty() {
-            anyhow::bail!("expected missing context references to be reported");
-        }
-        return Ok(());
-    }
-
-    if step.text == "line numbers should indicate error locations" {
-        // Line number tracking would be implemented in real streaming parser
-        return Ok(());
-    }
-
-    if step.text == "the handler should receive each fact" {
-        if world.streaming_context.facts_processed.is_empty() {
-            anyhow::bail!("handler did not receive any facts");
-        }
-        return Ok(());
-    }
-
-    if step.text == "contexts should be collected" {
-        if world.streaming_context.contexts_collected.is_empty() {
-            anyhow::bail!("no contexts were collected");
-        }
-        return Ok(());
-    }
-
-    if step.text == "units should be available for reference" {
-        if world.streaming_context.units_collected.is_empty() {
-            anyhow::bail!("no units were collected");
-        }
-        return Ok(());
-    }
-
-    // Taxonomy loader Then steps
-    if step.text == "the taxonomy should contain dimensions" {
-        let taxonomy = world
-            .taxonomy_loader_context
-            .taxonomy
-            .as_ref()
-            .context("taxonomy not loaded")?;
-        if taxonomy.dimensions.is_empty() {
-            anyhow::bail!("taxonomy has no dimensions");
-        }
-        return Ok(());
-    }
-
-    if step.text == "explicit dimensions should have domains" {
-        let taxonomy = world
-            .taxonomy_loader_context
-            .taxonomy
-            .as_ref()
-            .context("taxonomy not loaded")?;
-        let has_explicit_with_domain = taxonomy.dimensions.iter().any(|(_, d)| match d {
-            Dimension::Explicit { default_domain, .. } => default_domain.is_some(),
-            _ => false,
-        });
-        if !has_explicit_with_domain && !taxonomy.dimensions.is_empty() {
-            anyhow::bail!("no explicit dimensions have domains defined");
-        }
-        return Ok(());
-    }
-
-    if step.text == "domains should have members" {
-        let taxonomy = world
-            .taxonomy_loader_context
-            .taxonomy
-            .as_ref()
-            .context("taxonomy not loaded")?;
-        let has_members = taxonomy.domains.values().any(|d| !d.members.is_empty());
-        if !has_members {
-            anyhow::bail!("no domains have members defined");
-        }
-        return Ok(());
-    }
-
-    if step.text == "members should maintain parent-child relationships" {
-        return Ok(());
-    }
-
-    if step.text == "typed dimensions should have value types" {
-        return Ok(());
-    }
-
-    if step.text == "the value types should be valid XSD types" {
-        return Ok(());
-    }
-
-    if step.text == "hypercubes should contain their dimensions" {
-        return Ok(());
-    }
-
-    if step.text == "dimensions should reference their domains" {
-        let taxonomy = world
-            .taxonomy_loader_context
-            .taxonomy
-            .as_ref()
-            .context("taxonomy not loaded")?;
-        if taxonomy.dimension_domains.is_empty() && !taxonomy.dimensions.is_empty() {
-            anyhow::bail!("no dimension-domain references found");
-        }
-        return Ok(());
-    }
-
-    if step.text == "the taxonomy file should be cached" {
-        let cache_dir = world
-            .taxonomy_loader_context
-            .cache_dir
-            .as_ref()
-            .context("cache directory not configured")?;
-        if !cache_dir.exists() {
-            anyhow::bail!("cache directory does not exist");
-        }
-        return Ok(());
-    }
-
-    if step.text == "subsequent loads should use the cache" {
-        return Ok(());
-    }
-
-    if step.text == "imported schemas should be loaded" {
-        return Ok(());
-    }
-
-    if step.text == "all dimension definitions should be available" {
-        let taxonomy = world
-            .taxonomy_loader_context
-            .taxonomy
-            .as_ref()
-            .context("taxonomy not loaded")?;
-        if taxonomy.dimensions.is_empty() {
-            anyhow::bail!("no dimension definitions available");
-        }
-        return Ok(());
-    }
-
-    anyhow::bail!("unsupported BDD step: {}", step.text)
-}
-
-/// Create a synthetic taxonomy for testing when fixture files don't exist
-fn create_synthetic_taxonomy() -> DimensionTaxonomy {
-    let mut taxonomy = DimensionTaxonomy::new();
-
-    let mut scenario_domain = Domain::new("us-gaap:ScenarioDomain");
-    scenario_domain.add_member(DomainMember {
-        qname: "us-gaap:ScenarioActualMember".to_string(),
-        parent: None,
-        order: 1,
-        label: None,
-    });
-    scenario_domain.add_member(DomainMember {
-        qname: "us-gaap:ScenarioForecastMember".to_string(),
-        parent: None,
-        order: 2,
-        label: None,
-    });
-    taxonomy.add_domain(scenario_domain);
-
-    taxonomy.add_dimension(Dimension::Explicit {
-        qname: "us-gaap:StatementScenarioAxis".to_string(),
-        default_domain: Some("us-gaap:ScenarioDomain".to_string()),
-        required: false,
-    });
-    taxonomy.dimension_domains.insert(
-        "us-gaap:StatementScenarioAxis".to_string(),
-        "us-gaap:ScenarioDomain".to_string(),
-    );
-
-    taxonomy.add_dimension(Dimension::Typed {
-        qname: "dim:CustomerAxis".to_string(),
-        value_type: "xs:string".to_string(),
-        required: false,
-    });
-
-    taxonomy
-}
-
-fn parse_count_suffix(step: &str, prefix: &str, noun_stem: &str) -> Option<usize> {
-    let remainder = step.strip_prefix(prefix)?;
-    let count = remainder.split_whitespace().next()?.parse::<usize>().ok()?;
-    let noun = remainder
-        .split_whitespace()
-        .nth(1)
-        .unwrap_or_default()
-        .trim_end_matches('s');
-    if noun == noun_stem { Some(count) } else { None }
-}
-
-/// Select scenarios matching a selector (`scenario_id`, `ac_id`, `req_id`, or tag)
-fn select_matching_scenarios(grid: &FeatureGrid, selector: &str) -> Vec<ScenarioRecord> {
-    grid.scenarios
-        .iter()
-        .filter(|scenario| selector_matches(scenario, selector))
-        .cloned()
-        .collect()
-}
-
-/// Check if a scenario matches the given selector
-fn selector_matches(scenario: &ScenarioRecord, selector: &str) -> bool {
-    scenario.scenario_id == selector
-        || scenario.ac_id.as_deref() == Some(selector)
-        || scenario.req_id.as_deref() == Some(selector)
-        || format!("@{}", scenario.scenario_id) == selector
-        || scenario
-            .ac_id
-            .as_ref()
-            .is_some_and(|ac| format!("@{ac}") == selector)
-}
+//!
+//! This crate provides BDD step handlers for xbrlkit scenarios.
+//! The implementation is organized into submodules:
+//!
+//! - `types`: Core types including `World`, `Step`, and context structs
+//! - `given`: "Given" step handlers
+//! - `when`: "When" step handlers
+//! - `then`: "Then" step handlers and parameterized assertions
+//! - `utils`: Helper functions
+
+mod given;
+mod then;
+mod types;
+mod utils;
+mod when;
+
+pub use types::{Step, World, run_scenario};
